Remove LinuxPerf
We do not use perf tool anymore, let's remove it from the codebase.

Test: cmake --build .
