+++ OrbitCore/Params.h
@@ -32,7 +32,6 @@ struct Params {
   bool m_BpftraceCallstacks;
   bool m_SystemWideScheduling;
   bool m_UseBpftrace;
-  bool m_SampleWithPerf;
   int m_MaxNumTimers;
   float m_FontSize;
   int m_Port;
